refactor: error log 제거

+++ src/main/java/chess/service/ChessService.java
@@ -53,6 +53,9 @@ public class ChessService {
 
     public void deleteBy(Long roomId, String password) {
         Room room = roomDao.findById(roomId);
+        if (room.getStatus()) {
+            throw new IllegalArgumentException("진행 중인 게임은 삭제할 수 없습니다.");
+        }
         if (room.getPassword().equals(password)) {
             roomDao.deleteBy(roomId, password);
             return;
